refactor(api): unify CrossScopeEdge type and move to bunking domain layer
Addresses 5 review findings:

1. Move CrossScopeEdge Pydantic class from api/schemas/social_graph.py
   to bunking/graph/scope_filter.py to fix the layering inversion.
   Re-exported from api/schemas/social_graph.py for backward compat.

2. Unify the TypeScript wire type. Two interfaces described the same
   payload with different optionality. They now share a single named
   export CrossScopeEdge from frontend/src/types/graph.ts, reflecting
   Pydantic's actual wire shape. CrossScopeEdgeData kept as a type
   alias for callsite compatibility.

3. Update test_scope_filter.py to import CrossScopeEdge from the new
   domain location.

4. Restore section-header comments in api/schemas/__init__.py __all__
   for navigability; add noqa RUF022 since the rule wants strict
   alphabetical sort.

5. Update test fixture in cytoscapeStyles.test.ts to provide the now-
   required nullable fields explicitly.
